+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - [#1241](https://github.com/crypto-org-chain/cronos/pull/1241) Improve parallelization of memiavl restoration.
+- [#1302](https://github.com/crypto-org-chain/cronos/pull/1302) Fix concurrent map access in rootmulti store.
 
 ### State Machine Breaking